[diag] added base class for all diagnostics
