+++ veros/__init__.py
@@ -1,3 +1,3 @@
-from .veros_method import veros_method, veros_inline_method
+from .decorators import veros_method, veros_inline_method, veros_class_method
 from .veros import Veros
 from .veros_legacy import VerosLegacy
