+++ veros/diagnostics/energy.py
@@ -1,0 +1,14 @@
+from .diagnostic import VerosDiagnostic
+
+class Energy(VerosDiagnostic):
+    def initialize(self, veros):
+        import warnings
+        warnings.warn("routine is not implemented yet")
+
+    def diagnose(self,veros):
+        import warnings
+        warnings.warn("routine is not implemented yet")
+
+    def output(self,veros):
+        import warnings
+        warnings.warn("routine is not implemented yet")
