+++ veros/diagnostics/io_tools/__init__.py
@@ -1,0 +1,7 @@
+from . import netcdf
+
+threaded_io = netcdf.threaded_netcdf
+initialize_file = netcdf.initialize_netcdf_file
+add_dimension = netcdf.add_dimension
+initialize_variable = netcdf.initialize_variable
+write_variable = netcdf.write_variable
