--- veros/diagnostics_tools/__init__.py
@@ -1,1 +1,0 @@
-from . import averages, cfl_monitor, energy, overturning, particles, snapshot, tracer_monitor
