--- veros/diagnostics_tools/energy.py
@@ -1,19 +1,0 @@
-def init_diag_energy(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def diag_energy_read_restart(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def initialize(veros):
-    init_diag_energy(veros)
-    diag_energy_read_restart(veros)
-
-def diagnose(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def output(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
