--- veros/diagnostics_tools/io_tools/__init__.py
@@ -1,6 +1,0 @@
-from . import netcdf
-
-threaded_io = netcdf.threaded_netcdf
-initialize_file = netcdf.initialize_netcdf_file
-initialize_variable = netcdf.initialize_variable
-write_variable = netcdf.write_variable
