--- veros/diagnostics_tools/overturning.py
@@ -1,14 +1,0 @@
-def initialize(veros):
-    diag_over_read_restart(veros)
-
-def diag_over_read_restart(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def diagnose(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def output(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
