--- veros/diagnostics_tools/particles.py
@@ -1,25 +1,0 @@
-def initialize(veros):
-    set_particles(veros)
-    init_diag_particles(veros)
-    diag_particles_read_restart(veros)
-    diagnose(veros)
-
-def set_particles(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def init_diag_particles(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def diag_particles_read_restart(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def diagnose(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
-
-def output(veros):
-    import warnings
-    warnings.warn("routine is not implemented yet")
